modified neccesary files
